requirements.txt
asgiref==3.3.1
attrdict==2.0.1
base58==2.0.1
boto3==1.16.23
botocore==1.19.23
cytoolz==0.11.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-countries==6.1.3
django-crispy-forms==1.10.0
django-storages==1.10.1
ecdsa==0.16.1
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
gunicorn==20.0.4
hexbytes==0.2.1
jmespath==0.10.0
oauthlib==3.1.0
parsimonious==0.8.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pycryptodome==3.9.9
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests-oauthlib==1.3.0
rlp==1.2.0
s3transfer==0.3.3
sqlparse==0.4.1
stripe==2.55.0
toolz==0.11.1
tronapi==3.1.6
trx-utils==1.0.4
